c++ - Add variant visitor pattern accessor in c++ union class
In order to enforce at compile time that all members in a switch are handled.
As an add-on to the existing union accessors (disriminiator & individual getters)

Details:
To use it you define in c++ a "Visitor" struct with operator overloads taking each of the alternatives.

eg for a union in ADL:
union UnionTypeExample {
    String stringFieldA;
    String stringFieldB;
    Void voidFieldX;
    Void voidFieldY;
};

In c++ you can define a Visitor:

struct UnionTypeExampleVisitor {
    void stringFieldA(const std::string& stringFieldA) {
        // gets called in case of stringFieldA
    }

    void stringFieldB(const std::string& stringFieldB) {
        // gets called in case of stringFieldB
    }

    void voidFieldX() {
        // gets called in case of voidFieldX
    }

    void voidFieldY() {
        // gets called in case of voidFieldX
    }
};

And use it in c++ as follows:

UnionTypeExampleVisitor visitor;

UnionTypeExample unionInstance;
unionInstance.visit(visitor);

Based on boost::variant,  std::variant