c++ - Add variant visitor pattern accessor in c++ union class
In order to enforce at compile time that all members in a switch are handled.
As an add-on to the existing union accessors (disriminiator & individual getters)

Details:
To use it you define in c++ a "Visitor" struct with operator overloads taking each of the alternatives.
For each Void type alternative the visitor needs to have an overload accepting the matching std::integral_constant
For each non-void type alternative the visitor needs to have an overload accepting the matching member type and a matching std::integral_constant

eg for a union in ADL:
union UnionTypeExample {
    String stringFieldA;
    String stringFieldB;
    Void voidFieldX;
    Void voidFieldY;
};

In c++ you can define a Visitor:

struct UnionTypeExampleVisitor {
    void operator()( const std::string& stringFieldA, std::integral_constant<UnionTypeExample::DiscType, UnionTypeExample::DiscType::STRINGFIELDA>) {
        // gets called in case of stringFieldA
    }

    void operator()( const std::string& stringFieldB, std::integral_constant<UnionTypeExample::DiscType, UnionTypeExample::DiscType::STRINGFIELDB>) {
        // gets called in case of stringFieldB
    }

    void operator()( std::integral_constant<UnionTypeExample::DiscType, UnionTypeExample::DiscType::VOIDFIELDX>) {
        // gets called in case of voidFieldX
    }

    void operator()( std::integral_constant<UnionTypeExample::DiscType, UnionTypeExample::DiscType::VOIDFIELDY>) {
        // gets called in case of voidFieldX
    }
};

And use it in c++ as follows:

UnionTypeExampleVisitor visitor;

UnionTypeExample unionInstance;
unionInstance.visit(visitor);

Based on boost::variant,  std::variant